add login Function

+++ weevent-governance/src/main/java/com/webank/weevent/governance/entity/AccountExample.java
@@ -115,52 +115,52 @@ public class AccountExample {
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdEqualTo(Long value) {
+	        public Criteria andIdEqualTo(Integer value) {
 	            addCriterion("id =", value, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdNotEqualTo(Long value) {
+	        public Criteria andIdNotEqualTo(Integer value) {
 	            addCriterion("id <>", value, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdGreaterThan(Long value) {
+	        public Criteria andIdGreaterThan(Integer value) {
 	            addCriterion("id >", value, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdGreaterThanOrEqualTo(Long value) {
+	        public Criteria andIdGreaterThanOrEqualTo(Integer value) {
 	            addCriterion("id >=", value, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdLessThan(Long value) {
+	        public Criteria andIdLessThan(Integer value) {
 	            addCriterion("id <", value, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdLessThanOrEqualTo(Long value) {
+	        public Criteria andIdLessThanOrEqualTo(Integer value) {
 	            addCriterion("id <=", value, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdIn(List<Long> values) {
+	        public Criteria andIdIn(List<Integer> values) {
 	            addCriterion("id in", values, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdNotIn(List<Long> values) {
+	        public Criteria andIdNotIn(List<Integer> values) {
 	            addCriterion("id not in", values, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdBetween(Long value1, Long value2) {
+	        public Criteria andIdBetween(Integer value1, Integer value2) {
 	            addCriterion("id between", value1, value2, "id");
 	            return (Criteria) this;
 	        }
 
-	        public Criteria andIdNotBetween(Long value1, Long value2) {
+	        public Criteria andIdNotBetween(Integer value1, Integer value2) {
 	            addCriterion("id not between", value1, value2, "id");
 	            return (Criteria) this;
 	        }
